+++ client/components/CariSection/index.tsx
@@ -6,7 +6,7 @@ import TextInput from '@components/Form/TextInput';
 import Button from '@components/Button';
 import { mockgeoJSON } from '@lib/constants';
 import LocationList from '@components/LocationList';
-import { useGeocode } from 'context/geocode';
+import useGeocode from 'hooks/useGeocode';
 
 const CariSection = () => {
   const location = useLocation();
